test(voice-call): isolate media lifecycle coverage
